Add alternative layouts (fr, colemak)
